mrmevr: Output db
Switch access from mbbi/o to longin/out and use raw soft mbbi/o

The multiple mbbo's can be displayed as EDM choice buttons.
This depends on the readback behavior of this widget, that
if the readback differs then none of the buttons appear selected.

The mbbi's will read 'Illegal Value' when the mapping is not in range
which EDM displays does not display (shows empty string).