mrmevr: Keep the ISR from returning while the interrupt is being asserted
Fixes a race condition between acknowledging the VME IRQ and
the OS unmasking the interrupt.

The failure can occur when the ISR returns and the interrupt
has been unmasked before the card has stopped asserting the
interrupt.  This causes the VME bridge to attempt to service an
interrupt which the board is not asserting.  This causes a
bus error.

The "fix" is to read the interrupt flag again, this makes certain
that the card has at least one VME bus cycle to stop asserting
the interrupt.