On boot don't enable EVR until all configuration is set
All initialization happens during the initial phase (PINI=YES)
but the link state record will not start scanning until after
normal processing begins.  Thus on a cold boot the master enable
will not be turned on until after the board has been completely
configured.

During a warm boot the master enable will already be on.  In this
case trying to turn it on again will be a noop.