Generalize CardMap
CardMap is a way to associate several class instances with
a numeric global identifier.  This allows EPICS device support
and IOC shell functions to find instances.

There will be one instance of CardMap<INTRFACECLASS>
for each interface type exported by any drivers.
A potential user will query the instance to find
out if an ID number has been associated with
an instance of that interface.