mrmevr: avoid callback queue overflow
Ensure that CALLBACKs are not requested more often then they
can run.  Callbacks will try to run once for every time they are queued.
Which will result in an overflow if the time needed to run is longer
then the time between requests.

Because the callback API does not provide a mechanism to determine
if a callback is queued/run we take advantage of the fact that
callbacks are always executed in the order required.  By injecting
a special callback after other callbacks are requested we can know
when all the actions for one event are complete.

Conflicts:

	evrMrmApp/src/drvem.cpp