linux: quick fix for ISR race condition
There is a (unavoidable) race condition with the read-modify-write
of IRQEnable in the kernel ISR.

However, this shouldn't be a problem as the write only occurs if the
decision has been made to wake up the user-space ISR, which would then
handle/re-enable the interrupts.  Unless by chance, IRQEnable or IRQFlag
changes in that time so that "!active" is true, then interrupts are
never re-enabled.